requirements.txt
torch
transformers
scikit-learn
pillow
numpy
pandas
tqdm
matplotlib
opencv-python
